create festival prizes component
