+++ libs/movie/src/lib/movie/form/summary/festival-prizes/festival-prizes.component.scss
@@ -1,0 +1,3 @@
+.info-name {
+  opacity: 0.5;
+}
