+++ libs/movie/src/lib/movie/form/summary/festival-prizes/festival-prizes.component.ts
@@ -1,0 +1,12 @@
+import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
+import { MovieFestivalPrizesForm } from '@blockframes/movie/movieform/festival-prizes/festival-prizes.form';
+
+@Component({
+  selector: '[festivalPrizes] movie-summary-festival-prizes',
+  templateUrl: './festival-prizes.component.html',
+  styleUrls: ['./festival-prizes.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class MovieSummaryFestivalPrizesComponent {
+  @Input() festivalPrizes: MovieFestivalPrizesForm;
+}
