--- libs/movie/src/lib/movie/form/summary/main/festival-prizes/festival-prizes.component.html
@@ -1,118 +1,0 @@
-<h2>Title Information</h2>
-<div class="information" fxLayout="row" fxLayoutAlign="start center">
-  <h3>Main Information</h3>
-  <a mat-icon-button routerLink="../../../../../../main"><mat-icon svgIcon="edit"></mat-icon></a>
-</div>
-
-<p>
-  <span class="mat-body-1 info-name">Content Type </span>
-  <span class="mat-body-1">{{ workType | missing }}</span>
-</p>
-<p>
-  <span class="mat-body-1 info-name">Freshness </span>
-  <span class="mat-body-1">{{ storeType | missing }}</span>
-</p>
-<p>
-  <span class="mat-body-1 info-name">Production Status Mandatory </span>
-  <span class="mat-body-1">{{ status | missing }}</span>
-</p>
-<p>
-  <span class="mat-body-1 info-name">Original Title </span>
-  <span class="mat-body-1">{{ original | missing }}</span>
-</p>
-<p>
-  <span class="mat-body-1 info-name">International Title </span>
-  <span class="mat-body-1">{{ international | missing }}</span>
-</p>
-<p>
-  <span class="mat-body-1 info-name">Reference # </span>
-  <span class="mat-body-1">{{ reference | missing }}</span>
-</p>
-<ng-container *ngFor="let director of main.directors.controls; let isLast = last">
-  <p>
-    <span class="mat-body-1 info-name">Director Names </span>
-    <span class="mat-body-1">
-      {{ director.get('firstName') | missing }} {{ director.get('lastName') | missing }}
-    </span>
-  </p>
-  <p>
-    <span class="mat-body-1 info-name">Director's Filmography </span>
-    <span class="mat-body-1">{{ director.get('shortBiography') | missing }}</span>
-  </p>
-</ng-container>
-
-<ng-container *ngFor="let country of main.originCountries.controls; let i = index">
-  <div class="country">
-    <p class="mat-body-2">Countries of Origin #{{ i + 1 }}</p>
-    <p>
-      <span class="mat-body-1 info-name">Country </span>
-      <span class="mat-body-1">{{ country | missing }}</span>
-    </p>
-    <p>
-      <span class="mat-body-1 info-name">Domestic Distributor </span>
-      <span class="mat-body-1"></span>
-    </p>
-    <p>
-      <span class="mat-body-1 info-name">Release Media </span>
-      <ng-container *ngFor="let release of salesInfo.originalRelease.controls">
-        <span *ngIf="release.get('country').value === country.value" class="mat-body-1">
-          {{ release.get('media') | missing }}
-        </span>
-      </ng-container>
-    </p>
-    <p>
-      <span class="mat-body-1 info-name">Release Date </span>
-      <ng-container *ngFor="let release of salesInfo.originalRelease.controls">
-        <span *ngIf="release.get('country').value === country.value" class="mat-body-1">
-          {{ release.get('date') | missing }}
-        </span>
-      </ng-container>
-    </p>
-  </div>
-</ng-container>
-
-<p>
-  <span class="mat-body-1 info-name">Original Languages </span>
-  <ng-container *ngFor="let language of main.originalLanguages.controls; let isLast = last">
-    <span class="mat-body-1">{{ language | missing }}{{ isLast ? '' : ', ' }}</span>
-  </ng-container>
-</p>
-
-<p>
-  <span class="mat-body-1 info-name">Genres </span>
-  <ng-container *ngFor="let genre of genres; let isLast = last">
-    <span class="mat-body-1">{{ genre | missing }}{{ isLast ? '' : ', ' }}</span>
-  </ng-container>
-</p>
-
-<p>
-  <span class="mat-body-1 info-name">Running Time </span>
-  <span class="mat-body-1">{{ totalRunTime | missing }}</span>
-</p>
-
-<p>
-  <span class="mat-body-1 info-name">Principal Cast </span>
-  <ng-container *ngFor="let cast of salesCast.cast.controls; let isLast = last">
-    <span *ngIf="cast.value === 'lead-role'" class="mat-body-1">
-      {{ cast.get('firstName') | missing }} {{ cast.get('lastName') | missing }}{{ isLast ? '' : ', ' }}
-    </span>
-  </ng-container>
-</p>
-
-<p>
-  <span class="mat-body-1 info-name">Secondary Roles </span>
-  <ng-container *ngFor="let cast of salesCast.cast.controls; let isLast = last">
-    <span *ngIf="cast.value === 'secondary-role'" class="mat-body-1">
-      {{ cast.get('firstName') | missing }} {{ cast.get('lastName') | missing }}{{ isLast ? '' : ', ' }}
-    </span>
-  </ng-container>
-</p>
-
-<ng-container *ngFor="let festivalPrize of festivalPrizes.prizes.controls">
-  <p>
-    <span class="mat-body-1 info-name">Festival </span>
-    <span class="mat-body-1">
-      {{ festivalPrize.get('name') | missing }}, {{ festivalPrize.get('prize') | missing }}, {{ festivalPrize.get('year') | missing }}, {{ festivalPrize.get('premiere') | missing }}
-    </span>
-  </p>
-</ng-container>
