--- libs/movie/src/lib/movie/form/summary/main/festival-prizes/festival-prizes.component.scss
@@ -1,16 +1,0 @@
-.info-name {
-  opacity: 0.5;
-}
-
-.country {
-  margin-top: 16px;
-  margin-bottom: 16px;
-}
-
-h3 {
-  margin-bottom: 0;
-}
-
-.information {
-  margin-bottom: 20px;
-}
